LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/math-script-src.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/gen...

    <!-- speculative case in document.write -->
    <math><script src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=b2bdd18b-0c06-4429-8afb-851882be534b&amp;encodingcheck=&Gbreve;"></script></math>


FAIL Speculative parsing, document.write(): math-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/generated/document-write/math-script-src.tentative.sub.html"
